Fixes specs for resque v1.26.0
* Tests use normal fork and communicate with test runner via IO.pipe.
* Tests uses a zero interval to exit ASAP without timeout when done.
* Children need to `exit!` to avoid running test/unit at_exit hooks.
* Gemspec depends on a specific minor version of resque.

We must depend on ~> v1.26.0, because these methods are considered
internal, are not a stable API, and can change with every version. I.e.
as written, the tests do not pass with v1.25.0 nor with v1.27.0. Until
Resque::Worker has a stable API we can target, this gem will need to be
tested with each minor release.