Fix flaky TERM_CHILD support. Switch TSTP signal to QUIT.
In tests, `SIGTSTP` is often ignored for ruby 2.3-2.5 (usually works in
2.6?). I assume this has something to do with some signals temporarily
being blocked by ruby, but I'm really not sure.  However, in my tests,
`SIGQUIT` always gets through to the handler.  Since `SIGQUIT` is
normally registered to call `shutdown` anyway, I've simply re-registered
it to continue that role in the `term_child` environment.