Use `log_with_severity` over deprecated `log`/`log!`
The Resque gem stopped using the `log` and `log!` methods internally in
the `Resque::Worker` class some time ago and replaced them with the
`log_with_severity` method (https://github.com/resque/resque/pull/1405).

Although the `log` and `log!` methods still exist for backwards
compatibility, this change switches to using the more recent interface
to stay consistent with Resque's internal style.

(The `log_with_severity` method was introduced with resque v1.26.0,
which is within the `>= 1.27.0", "< 2.1"` compatibility range in the
gemspec.)