Avoid discarding jobs on SIGTERM
Previously, a child receiving TERM would immediately raise an
exception. If this happened during or after Resque.reserve but
before the job started, the exception would not be handled like a
job failure. As a result, resque-retry never sees the failure and
the job is lost permanently.

A more complete solution for this would probably require rewriting
parts of Resque itself to use LMOVE rather than LPOP, but this
should bring the risk of job loss on par with vanilla Resque.